ml/requirements.txt
pandas
numpy
scikit-learn
transformers
torch
grpcio
grpcio-tools